Thread slide show
my new idea for handling the slide show will be to use a thread. As long as the amount of time per slide is < the time that "get ready is displayed. "get ready" will display over the last slide and the loop in slide show will end during the first for loop when the READY_LED is flashing. Because the program is doing nothing but waiting for a button to be pushed the threading should not bog down the CPU too much.